auth/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
typing_extensions==4.4.0
